src/main/java/com/safetyNet/safetyNetAlerts/service/MedicalRecordService.java
package com.safetyNet.safetyNetAlerts.service;

import com.safetyNet.safetyNetAlerts.model.MedicalRecord;
import com.safetyNet.safetyNetAlerts.repository.MedicalRecordRepository;

import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class MedicalRecordService {
    
    private final MedicalRecordRepository medicalRecordRepository;

    public MedicalRecordService(MedicalRecordRepository medicalRecordRepository) {
        this.medicalRecordRepository = medicalRecordRepository;
    }

    public List<MedicalRecord> getAllMedicalRecords() {
        return medicalRecordRepository.getAllMedicalRecords();
    }

    public Optional<MedicalRecord> getMedicalRecordByName(String firstName, String lastName) {
        return medicalRecordRepository.getMedicalRecordByName(firstName, lastName);
    }

    public MedicalRecord addMedicalRecord(MedicalRecord medicalRecord) {
        return medicalRecordRepository.addMedicalRecord(medicalRecord);
    }

    public void deleteMedicalRecord(String firstName, String lastName) {
        medicalRecordRepository.deleteMedicalRecord(firstName, lastName);
    }

    public MedicalRecord updateMedicalRecord(MedicalRecord existingMedicalRecord, MedicalRecord updatedMedicalRecord) {
        return medicalRecordRepository.updateMedicalRecord(existingMedicalRecord, updatedMedicalRecord);
    }
}
